REST_create_object:
Use javascript/JQuery/AJax and REST API to create object, without having
to refresh the entire page... but it is a problem to add edit function
to the page.

So, commiting this as a branch... will fallback to the previous
approach: use view + refresh page to add or edit object.